move data resolving to own file
